refactor: 채팅 불필요한 전역상태 제거
