--- src/components/room/chatRoom/InitMessages.tsx
@@ -1,16 +1,0 @@
-'use client';
-import { IMessage, useMessageStore } from '@/store/messageStore';
-import { useEffect, useRef } from 'react';
-
-export const InitMessages = ({ messages }: { messages: IMessage[] }) => {
-  const initState = useRef(false);
-
-  useEffect(() => {
-    if (!initState.current) {
-      useMessageStore.setState({ messages });
-    }
-    initState.current = true;
-  }, []);
-
-  return <></>;
-};
